requirements.txt
streamlitlibrosanumpymatplotlibstreamlit-mic-recorderscipy